+++ src/app/admin/page.tsx
@@ -1,249 +1,12 @@
 "use client";
 
-import { useState } from "react";
-import AdminSidebar from "./components/admin-sidebar";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Edit, Trash2, Plus } from "lucide-react";
+import { useEffect } from "react";
 
-export default function AdminDashboard() {
-  console.log("Admin dashboard component rendered");
-  const [products, setProducts] = useState([
-    {
-      id: 1,
-      name: "Premium Headphones",
-      price: 199.99,
-      stock: 45,
-      category: "Electronics",
-    },
-    {
-      id: 2,
-      name: "Wireless Keyboard",
-      price: 59.99,
-      stock: 32,
-      category: "Electronics",
-    },
-    {
-      id: 3,
-      name: "Ergonomic Chair",
-      price: 249.99,
-      stock: 15,
-      category: "Furniture",
-    },
-    {
-      id: 4,
-      name: "Smart Watch",
-      price: 149.99,
-      stock: 28,
-      category: "Electronics",
-    },
-    {
-      id: 5,
-      name: "Laptop Backpack",
-      price: 79.99,
-      stock: 53,
-      category: "Accessories",
-    },
-  ]);
+export default function AdminPage() {
+  useEffect(() => {
+    // Redirect to admin dashboard immediately
+    window.location.href = "/admin/login";
+  }, []);
 
-  const [categories, setCategories] = useState([
-    { id: 1, name: "Electronics", productCount: 15 },
-    { id: 2, name: "Clothing", productCount: 24 },
-    { id: 3, name: "Home & Kitchen", productCount: 18 },
-    { id: 4, name: "Books", productCount: 32 },
-    { id: 5, name: "Furniture", productCount: 9 },
-    { id: 6, name: "Accessories", productCount: 21 },
-  ]);
-
-  return (
-    <div className="min-h-screen bg-gray-50">
-      <div className="flex flex-col md:flex-row">
-        <AdminSidebar activePage="products" />
-
-        <div className="flex-1 p-6 md:p-8 md:ml-64">
-          <Tabs defaultValue="products" className="w-full">
-            <TabsList className="mb-6">
-              <TabsTrigger value="products">Products</TabsTrigger>
-              <TabsTrigger value="categories">Categories</TabsTrigger>
-              <TabsTrigger value="analytics">Analytics</TabsTrigger>
-            </TabsList>
-
-            <TabsContent value="products" className="space-y-6">
-              <div className="flex justify-between items-center">
-                <h2 className="text-2xl font-bold">Products</h2>
-                <button className="bg-blue-600 text-white px-4 py-2 rounded-lg flex items-center gap-2 hover:bg-blue-700 transition-colors">
-                  <Plus size={18} />
-                  <span>Add Product</span>
-                </button>
-              </div>
-
-              <div className="bg-white shadow rounded-lg overflow-hidden">
-                <div className="overflow-x-auto">
-                  <table className="min-w-full divide-y divide-gray-200">
-                    <thead className="bg-gray-50">
-                      <tr>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Product Name
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Price
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Stock
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Category
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Actions
-                        </th>
-                      </tr>
-                    </thead>
-                    <tbody className="divide-y divide-gray-200">
-                      {products.map((product) => (
-                        <tr key={product.id}>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="text-sm font-medium text-gray-900">
-                              {product.name}
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="text-sm text-gray-900">
-                              ${product.price.toFixed(2)}
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="text-sm text-gray-900">
-                              {product.stock}
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="text-sm text-gray-900">
-                              {product.category}
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap text-sm font-medium">
-                            <div className="flex gap-2">
-                              <button className="text-blue-600 hover:text-blue-900">
-                                <Edit size={18} />
-                              </button>
-                              <button className="text-red-600 hover:text-red-900">
-                                <Trash2 size={18} />
-                              </button>
-                            </div>
-                          </td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            </TabsContent>
-
-            <TabsContent value="categories" className="space-y-6">
-              <div className="flex justify-between items-center">
-                <h2 className="text-2xl font-bold">Categories</h2>
-                <button className="bg-blue-600 text-white px-4 py-2 rounded-lg flex items-center gap-2 hover:bg-blue-700 transition-colors">
-                  <Plus size={18} />
-                  <span>Add Category</span>
-                </button>
-              </div>
-
-              <div className="bg-white shadow rounded-lg overflow-hidden">
-                <div className="overflow-x-auto">
-                  <table className="min-w-full divide-y divide-gray-200">
-                    <thead className="bg-gray-50">
-                      <tr>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Category Name
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Products
-                        </th>
-                        <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                          Actions
-                        </th>
-                      </tr>
-                    </thead>
-                    <tbody className="divide-y divide-gray-200">
-                      {categories.map((category) => (
-                        <tr key={category.id}>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="text-sm font-medium text-gray-900">
-                              {category.name}
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap">
-                            <div className="text-sm text-gray-900">
-                              {category.productCount}
-                            </div>
-                          </td>
-                          <td className="px-6 py-4 whitespace-nowrap text-sm font-medium">
-                            <div className="flex gap-2">
-                              <button className="text-blue-600 hover:text-blue-900">
-                                <Edit size={18} />
-                              </button>
-                              <button className="text-red-600 hover:text-red-900">
-                                <Trash2 size={18} />
-                              </button>
-                            </div>
-                          </td>
-                        </tr>
-                      ))}
-                    </tbody>
-                  </table>
-                </div>
-              </div>
-            </TabsContent>
-
-            <TabsContent value="analytics" className="space-y-6">
-              <h2 className="text-2xl font-bold">Analytics</h2>
-
-              <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-                <div className="bg-white p-6 rounded-lg shadow">
-                  <h3 className="text-lg font-medium text-gray-900 mb-2">
-                    Total Sales
-                  </h3>
-                  <p className="text-3xl font-bold">$12,345</p>
-                  <p className="text-sm text-green-600 mt-2">
-                    +12% from last month
-                  </p>
-                </div>
-
-                <div className="bg-white p-6 rounded-lg shadow">
-                  <h3 className="text-lg font-medium text-gray-900 mb-2">
-                    Orders
-                  </h3>
-                  <p className="text-3xl font-bold">156</p>
-                  <p className="text-sm text-green-600 mt-2">
-                    +8% from last month
-                  </p>
-                </div>
-
-                <div className="bg-white p-6 rounded-lg shadow">
-                  <h3 className="text-lg font-medium text-gray-900 mb-2">
-                    Customers
-                  </h3>
-                  <p className="text-3xl font-bold">86</p>
-                  <p className="text-sm text-green-600 mt-2">
-                    +15% from last month
-                  </p>
-                </div>
-              </div>
-
-              <div className="bg-white p-6 rounded-lg shadow">
-                <h3 className="text-lg font-medium text-gray-900 mb-4">
-                  Sales Overview
-                </h3>
-                <div className="h-64 flex items-center justify-center">
-                  <p className="text-gray-500">
-                    Sales chart will be displayed here
-                  </p>
-                </div>
-              </div>
-            </TabsContent>
-          </Tabs>
-        </div>
-      </div>
-    </div>
-  );
+  return null; // No need to render anything as we're redirecting
 }
